Progress on enemy AI
General state switching is working, but tracking logic is bugged. Inefficient calling of Coroutines in Update causes tracking to happen constantly, so the enemy never stops tracking the player. UpdateLastSeenPosition coroutine needs to be relocated in a more efficient spot - hopefully once that's done the skeleton of the system will be in place properly and a testable prototype can be devised